+++ components/table/index.tsx
@@ -6,7 +6,7 @@ import { Table as ChakraTable, Tbody, Th, Thead, Tr } from "@chakra-ui/table";
 import Image from "next/image";
 import BigTableRow from "./BigTableRow";
 import SmallTableRow from "./SmallTableRow";
-import useChakraBreakpoints from "@/utils/useChakraBreakpoints";
+import useChakraBreakpoints from "@/hooks/useChakraBreakpoints";
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const { useTable, usePagination } = require("react-table");
